third_party/blink/web_tests/external/wpt/css/mediaqueries/prefers-reduced-transparency-expected.txt

This is a testharness.js-based test.
FAIL Should be known: '(prefers-reduced-transparency)' assert_true: expected true got false
FAIL Should be known: '(prefers-reduced-transparency: no-preference)' assert_true: expected true got false
FAIL Should be known: '(prefers-reduced-transparency: reduce)' assert_true: expected true got false
PASS Should be parseable: '(prefers-reduced-transparency: 0)'
PASS Should be unknown: '(prefers-reduced-transparency: 0)'
PASS Should be parseable: '(prefers-reduced-transparency: none)'
PASS Should be unknown: '(prefers-reduced-transparency: none)'
PASS Should be parseable: '(prefers-reduced-transparency: 10px)'
PASS Should be unknown: '(prefers-reduced-transparency: 10px)'
PASS Should be parseable: '(prefers-reduced-transparency: no-preference reduce)'
PASS Should be unknown: '(prefers-reduced-transparency: no-preference reduce)'
PASS Should be parseable: '(prefers-reduced-transparency: reduced)'
PASS Should be unknown: '(prefers-reduced-transparency: reduced)'
PASS Should be parseable: '(prefers-reduced-transparency: no-preference/reduce)'
PASS Should be unknown: '(prefers-reduced-transparency: no-preference/reduce)'
FAIL Check that no-preference evaluates to false in the boolean context assert_equals: expected true but got false
PASS Check that invalid evaluates to false
Harness: the test ran to completion.
